\name{PVCurve-package}
\alias{PVCurve-package}
\alias{PVCurve}
\docType{package}
\title{
\packageTitle{PVCurve}
}
\description{
\packageDescription{PVCurve}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PVCurve}
\packageIndices{PVCurve}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{PVCurve}

Maintainer: \packageMaintainer{PVCurve}
}
\keyword{ package }
